fix(tools): audit-driven correctness batch + full tool-surface audit
Per the 17-agent tool-surface audit (docs/TOOL-AUDIT.md, avg quality 3.96/5,
18 buggy tools found):

- doc_perspective_guide: fix vp ray cast — (1.0 - t.fract()) was truncated to
  i32 before the x width, collapsing every bottom-edge ray to x=0.
- doc_move_region: paste source-over so transparent pixels in the moved block
  no longer punch a rectangular hole through art at the destination.
- doc_select_wand: flood no longer crosses the opaque/transparent boundary
  (oklab_delta ignores alpha, so a wand on a black fill bled into a black
  transparent background).
- doc_create: reject 0-px and >4096 canvases instead of accepting 0/OOM dims.
- doc_add_layer / doc_set_layer: reject unknown blend strings (silently
  composited Normal before) with the valid list.
- doc_add_tag: reject unknown direction (silently played forward).
- doc_add_frame: reject out-of-range copy_from (was a silent blank frame).
- doc_transform_cel: snap_palette now defaults true (parity with
  keyframe_transform; only acts when a palette is locked).
- palette_ramp: build on the OKLCh make_ramp_oklch engine (was linear-HSL that
  crushed midtones to near-black mud).

Adds docs/TOOL-AUDIT.md (full 103-tool audit + adversarial critic). 182 lib
tests (+5), clippy clean.

+++ src/studio/craft.rs
@@ -1063,7 +1063,7 @@ impl Studio {
                             let (ex, ey) = match t as i32 {
                                 0 => ((t.fract() * w as f32) as i32, 0),
                                 1 => (w - 1, (t.fract() * h as f32) as i32),
-                                2 => ((1.0 - t.fract()) as i32 * (w - 1), h - 1),
+                                2 => (((1.0 - t.fract()) * (w - 1) as f32) as i32, h - 1),
                                 _ => (0, ((1.0 - t.fract()) * h as f32) as i32),
                             };
                             d.line(li, f, vx, vy, ex, ey, color, 1)?;
